gemini_image_generation/requirements.txt
langchain
langchain_core
langchain-openai
dotenv
langchain_google_genai
langchain_community
pillow